Finally exact pathfinding. Might fix last-step-lag later, but its barely noticable so ill just finish the vladE-Ability now

+++ Player.hpp
@@ -14,7 +14,7 @@ public:
     int* pathXpositions;
     int* pathYpositions;
     int pathLenght;
-    float cPathIndex;
+    int cPathIndex;
 private:
 
 	int row;
